# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopClone E-Commerce API",
    "description": "A full-featured e-commerce REST API powering the ShopClone storefront. Provides endpoints for user authentication, product catalog browsing, shopping cart management, order processing, wishlists, product reviews, and simulated payment processing.",
    "version": "1.0.0",
    "contact": {
      "name": "ShopClone API Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Creates a new user account with the provided name, email, date of birth, and password. Returns the created user object and an authentication token.",
        "operationId": "signUp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              },
              "example": {
                "name": "Jane Doe",
                "email": "jane.doe@example.com",
                "dob": "1995-06-15",
                "password": "SecureP@ss123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "user": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "Jane Doe",
                    "email": "jane.doe@example.com",
                    "dob": "1995-06-15",
                    "address": {
                      "street": "",
                      "city": "",
                      "state": "",
                      "zipCode": "",
                      "country": ""
                    }
                  },
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body (missing or malformed fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Name, email, date of birth, and password are required"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "A user with this email already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in an existing user",
        "description": "Authenticates a user with their email and password. Returns the user object and a Bearer token for subsequent authenticated requests.",
        "operationId": "signIn",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              },
              "example": {
                "email": "jane.doe@example.com",
                "password": "SecureP@ss123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "user": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "Jane Doe",
                    "email": "jane.doe@example.com",
                    "dob": "1995-06-15",
                    "address": {
                      "street": "",
                      "city": "",
                      "state": "",
                      "zipCode": "",
                      "country": ""
                    }
                  },
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body (missing email or password)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Email and password are required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid email or password"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/auth/profile": {
      "put": {
        "tags": ["Auth"],
        "summary": "Update user profile",
        "description": "Updates the authenticated user's profile. All fields are optional; only provided fields will be updated.",
        "operationId": "updateProfile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              },
              "example": {
                "name": "Jane Smith",
                "email": "jane.smith@example.com",
                "dob": "1995-06-15",
                "address": {
                  "street": "123 Main Street, Apt 4B",
                  "city": "San Francisco",
                  "state": "CA",
                  "zipCode": "94102",
                  "country": "United States"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "user": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "Jane Smith",
                    "email": "jane.smith@example.com",
                    "dob": "1995-06-15",
                    "address": {
                      "street": "123 Main Street, Apt 4B",
                      "city": "San Francisco",
                      "state": "CA",
                      "zipCode": "94102",
                      "country": "United States"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Email already in use by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "A user with this email already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/auth/password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Change user password",
        "description": "Changes the authenticated user's password. Requires the current password for verification and the new password.",
        "operationId": "changePassword",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              },
              "example": {
                "currentPassword": "SecureP@ss123",
                "newPassword": "EvenMoreSecure@456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password successfully changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing fields or new password does not meet requirements)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Current password and new password are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/auth/account": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Delete user account",
        "description": "Permanently deletes the authenticated user's account. Requires the current password for confirmation.",
        "operationId": "deleteAccount",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              },
              "example": {
                "password": "SecureP@ss123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing password)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Password is required to delete account"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/cart/{productId}": {
      "delete": {
        "tags": ["Cart"],
        "summary": "Remove a specific item from the cart",
        "description": "Removes a single product from the authenticated user's shopping cart by its product ID.",
        "operationId": "removeCartItem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the product to remove from the cart",
            "schema": {
              "type": "string"
            },
            "example": "prod-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Item successfully removed; returns the updated cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                },
                "example": {
                  "items": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid product ID"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the /api/auth/signup or /api/auth/signin endpoints. Include in the Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
      "Address": {
        "type": "object",
        "description": "Represents a user's shipping address",
        "properties": {
          "street": {
            "type": "string",
            "description": "Street address",
            "example": "123 Main Street, Apt 4B"
          },
          "city": {
            "type": "string",
            "description": "City name",
            "example": "San Francisco"
          },
          "state": {
            "type": "string",
            "description": "State or region",
            "example": "CA"
          },
          "zipCode": {
            "type": "string",
            "description": "ZIP or postal code",
            "example": "94102"
          },
          "country": {
            "type": "string",
            "description": "Country name",
            "example": "United States"
          }
        }
      },
      "User": {
        "type": "object",
        "description": "Represents a user in the system",
        "required": ["id", "name", "email", "dob"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the user",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "jane.doe@example.com"
          },
          "dob": {
            "type": "string",
            "description": "Date of birth of the user",
            "example": "1995-06-15"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          }
        }
      },
      "Product": {
        "type": "object",
        "description": "Represents a product in the catalog",
        "required": ["id", "title", "description", "price", "image", "category"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the product",
            "example": "prod-001"
          },
          "title": {
            "type": "string",
            "description": "Name of the product",
            "example": "Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the product",
            "example": "Premium noise-cancelling wireless headphones with 30-hour battery life."
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Price of the product in USD",
            "minimum": 0,
            "example": 79.99
          },
          "image": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/images/headphones.jpg"
          },
          "category": {
            "type": "string",
            "description": "Category the product belongs to",
            "example": "electronics"
          }
        }
      },
      "CartItem": {
        "type": "object",
        "description": "Represents an item in the shopping cart, combining a product with a quantity",
        "required": ["product", "quantity"],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity of the product in the cart",
            "minimum": 1,
            "example": 2
          }
        }
      },
...
      "SignUpRequest": {
        "type": "object",
        "description": "Request body for user registration",
        "required": ["name", "email", "dob", "password"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "minLength": 1,
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address for the new account",
            "example": "jane.doe@example.com"
          },
          "dob": {
            "type": "string",
            "description": "Date of birth (YYYY-MM-DD format)",
            "example": "1995-06-15"
          },
          "password": {
            "type": "string",
            "description": "Password for the new account",
            "minLength": 6,
            "format": "password",
            "example": "SecureP@ss123"
          }
        }
      },
      "SignInRequest": {
        "type": "object",
        "description": "Request body for user authentication",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Registered email address",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Account password",
            "format": "password",
            "example": "SecureP@ss123"
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "description": "Request body for updating user profile. All fields are optional; only provided fields will be updated.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated full name",
            "minLength": 1,
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Updated email address",
            "example": "jane.smith@example.com"
          },
          "dob": {
            "type": "string",
            "description": "Updated date of birth (YYYY-MM-DD format)",
            "example": "1995-06-15"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "description": "Request body for changing the user's password",
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "The user's current password for verification",
            "format": "password",
            "example": "SecureP@ss123"
          },
          "newPassword": {
            "type": "string",
            "description": "The desired new password",
            "minLength": 6,
            "format": "password",
            "example": "EvenMoreSecure@456"
          }
        }
      },
      "DeleteAccountRequest": {
        "type": "object",
        "description": "Request body for account deletion confirmation",
        "required": ["password"],
        "properties": {
          "password": {
            "type": "string",
            "description": "Current password to confirm account deletion",
            "format": "password",
            "example": "SecureP@ss123"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "description": "Response returned after successful signup or signin",
        "required": ["user", "token"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "description": "JWT authentication token to be used as a Bearer token in subsequent requests",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "description": "Response containing the user profile",
        "required": ["user"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "description": "Generic success response",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates the operation was successful",
            "example": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message describing what went wrong",
            "example": "An unexpected error occurred"
          }
        }
      },
...
      "CartResponse": {
        "type": "object",
        "description": "Response containing the current state of the shopping cart",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "description": "Array of cart items (product with quantity)",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token is invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Authentication required"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}